2.11.4 release

+++ play.it-2/games/play-dark-reign-2.sh
@@ -1,8 +1,8 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
-# Copyright (c) 2015-2018, Antoine Le Gonidec
+# Copyright (c) 2015-2020, Antoine "vv221/vv222" Le Gonidec
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,18 +30,16 @@ set -o errexit
 
 ###
 # Dark Reign 2
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200301.2
 
 # Set game-specific variables
 
 GAME_ID='dark-reign-2'
 GAME_NAME='Dark Reign 2'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='setup_dark_reign2_2.0.0.11.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/dark_reign_2'
@@ -49,26 +47,25 @@ ARCHIVE_GOG_MD5='9a3d10825507b73c4db178f9caea2406'
 ARCHIVE_GOG_VERSION='1.3.882-gog2.0.0.11'
 ARCHIVE_GOG_SIZE='450000'
 
-ARCHIVE_DOC_PATH='app'
-ARCHIVE_DOC_FILES='./customer_support.htm ./manual.pdf ./readme.rtf ./*.txt'
+ARCHIVE_DOC_DATA_PATH='app'
+ARCHIVE_DOC_DATA_FILES='*.htm *.pdf *.rtf *.txt'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./dr2.exe ./launcher.exe ./binkw32.dll ./getinfo.dll ./libogg-0.dll ./libvorbis-0.dll ./libvorbisfile-3.dll ./mss32.dll ./msvcp90.dll ./msvcr90.dll ./winmm.dll ./anet.inf ./library'
+ARCHIVE_GAME_BIN_FILES='*.dll *.exe *.inf library'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./missions ./mods ./music ./packs ./sides ./worlds'
+ARCHIVE_GAME_DATA_FILES='missions mods music packs sides worlds'
 
 CONFIG_FILES='./settings.cfg'
 DATA_DIRS='./mods ./users'
 DATA_FILES='./dr2.log'
 
+APP_REGEDIT='dr2-cdkey.reg'
 APP_WINETRICKS='win98'
 
 APP_MAIN_TYPE='wine'
-APP_MAIN_PRERUN='regedit dr2-cdkey.reg'
 APP_MAIN_EXE='launcher.exe'
 APP_MAIN_ICON='dr2.exe'
-APP_MAIN_ICON_RES='16 32'
 
 PACKAGES_LIST='PKG_DATA PKG_BIN'
 
@@ -76,54 +73,62 @@ PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_DATA_ID, winetricks, wine32-development | wine32 | wine-bin | wine-i386 | wine-staging-i386, wine:amd64 | wine"
-PKG_BIN_DEPS_ARCH="$PKG_DATA_ID winetricks wine"
+PKG_BIN_DEPS="$PKG_DATA_ID wine winetricks glx"
 
 # Load common functions
 
-target_version='2.0'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
-
-PKG='PKG_DATA'
-organize_data 'GAME_DATA' "$PATH_GAME"
-organize_data 'DOC'       "$PATH_DOC"
-
-PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
-cat > "${PKG_BIN_PATH}${PATH_GAME}/dr2-cdkey.reg" << EOF
-REGEDIT4
-[HKEY_LOCAL_MACHINE\Software\WON\CDKeys]
-"DarkReign2"=hex:56,c1,0c,ed,bb,61,40,19,99,3d,cd,6c,78,51,4c,5e
-EOF
-
+prepare_package_layout
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Get game icon
+
+PKG='PKG_BIN'
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
+
+# Work around prompt for CD-key
+
+if [ "$DRY_RUN" -eq 0 ]; then
+	cat > "${PKG_BIN_PATH}${PATH_GAME}/$APP_REGEDIT" <<- 'EOF'
+	REGEDIT4
+	[HKEY_LOCAL_MACHINE\Software\WON\CDKeys]
+	"DarkReign2"=hex:56,c1,0c,ed,bb,61,40,19,99,3d,cd,6c,78,51,4c,5e
+	EOF
+fi
 
 # Build package
