+++ play.it-2/games/play-kings-bounty-the-legend.sh
@@ -1,8 +1,8 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
-# Copyright (c) 2015-2018, Antoine Le Gonidec
+# Copyright (c) 2015-2020, Antoine "vv221/vv222" Le Gonidec
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,23 +30,22 @@ set -o errexit
 
 ###
 # King’s Bounty: The Legend
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200229.1
 
 # Set game-specific variables
 
 GAME_ID='kings-bounty-the-legend'
-GAME_NAME='King’s Bounty: The Legend'
+GAME_NAME='Kingʼs Bounty: The Legend'
 
 ARCHIVES_LIST='ARCHIVE_GOG_EN ARCHIVE_GOG_FR'
 
 ARCHIVE_GOG_EN='setup_kings_bounty_the_legend_1.7_(15542).exe'
 ARCHIVE_GOG_EN_URL='https://www.gog.com/game/kings_bounty_the_legend'
 ARCHIVE_GOG_EN_MD5='f7a9defe0fd96a7f8d6dff6ed7828242'
-ARCHIVE_GOG_EN_TYPE='innosetup'
 ARCHIVE_GOG_EN_VERSION='1.7-gog15542'
 ARCHIVE_GOG_EN_SIZE='6000000'
 ARCHIVE_GOG_EN_PART1='setup_kings_bounty_the_legend_1.7_(15542)-1.bin'
@@ -56,7 +55,6 @@ ARCHIVE_GOG_EN_PART1_TYPE='innosetup'
 ARCHIVE_GOG_FR='setup_kings_bounty_the_legend_french_1.7_(15542).exe'
 ARCHIVE_GOG_FR_URL='https://www.gog.com/game/kings_bounty_the_legend'
 ARCHIVE_GOG_FR_MD5='646fdfacadc498826be127fe6703f259'
-ARCHIVE_GOG_FR_TYPE='innosetup'
 ARCHIVE_GOG_FR_VERSION='1.7-gog15542'
 ARCHIVE_GOG_FR_SIZE='6000000'
 ARCHIVE_GOG_FR_PART1='setup_kings_bounty_the_legend_french_1.7_(15542)-1.bin'
@@ -64,16 +62,16 @@ ARCHIVE_GOG_FR_PART1_MD5='907882679fb7050e172994d36730454a'
 ARCHIVE_GOG_FR_PART1_TYPE='innosetup'
 
 ARCHIVE_DOC_L10N_PATH='app'
-ARCHIVE_DOC_L10N_FILES='./readme.rtf ./manual.pdf'
+ARCHIVE_DOC_L10N_FILES='readme.rtf manual.pdf'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./*.dll ./kb.exe ./data/*.ini ./data/fonts.cfg'
+ARCHIVE_GAME_BIN_FILES='*.dll kb.exe data/default.ini data/game.ini data/fonts.cfg'
 
 ARCHIVE_GAME_L10N_PATH='app'
-ARCHIVE_GAME_L10N_FILES='./data/app.ini ./data/loc_data.kfs ./sessions/base/loc_ses.kfs'
+ARCHIVE_GAME_L10N_FILES='data/app.ini data/loc_data.kfs sessions/base/loc_ses.kfs'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./curver.txt ./data/animation.kfs ./data/data.kfs ./data/interface_textures.kfs ./data/models.kfs ./data/sky.kfs ./data/sounds.kfs ./data/textures.kfs ./data/calibri.ttf ./data/music ./data/video ./sessions/base/ses.kfs ./sessions/base/locations'
+ARCHIVE_GAME_DATA_FILES='curver.txt data sessions'
 
 CONFIG_FILES='./data/*.ini ./data/fonts.cfg'
 
@@ -82,9 +80,8 @@ APP_WINETRICKS='d3dx9'
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='kb.exe'
 APP_MAIN_ICON='kb.exe'
-APP_MAIN_ICON_RES='16 24 32 48'
 
-PACKAGES_LIST='PKG_L10N PKG_DATA PKG_BIN'
+PACKAGES_LIST='PKG_BIN PKG_L10N PKG_DATA'
 
 PKG_L10N_ID="${GAME_ID}-l10n"
 PKG_L10N_ID_GOG_EN="${PKG_L10N_ID}-en"
@@ -101,48 +98,48 @@ PKG_BIN_DEPS="$PKG_L10N_ID $PKG_DATA_ID wine"
 
 # Load common functions
 
-target_version='2.3'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
-
-# Check that all parts of the installer are present
-
-ARCHIVE_MAIN="$ARCHIVE"
-set_archive 'ARCHIVE_PART1' "${ARCHIVE_MAIN}_PART1"
-[ "$ARCHIVE_PART1" ] || set_archive_error_not_found "${ARCHIVE_MAIN}_PART1"
-ARCHIVE="$ARCHIVE_MAIN"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"   "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}"  "$PATH_GAME"
-done
+# Get game icon
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
